refactor: split the renderer into a package
renderer.py becomes paz/graphics/renderer/, one module per stage of the
pipeline: rays (the bounce loop), tiling, intersect, shade, shadow,
optics, material. The loop body now reads intersect -> shade -> advance
instead of burying the scatter step inside update_state.

Avoids "trace", which in a JAX-first library already means jit tracing:
trace_bounces is rays.render, trace_chunks is rays.render_chunks.

The four shape-rendering examples patched
paz.graphics.renderer.compute_soft_occlusion, which the split would have
silently turned into a no-op; they now patch renderer.shadow directly.
The tiling module is not called "tiles" because render() takes a public
tiles= keyword that examples pass by name.

No behaviour change. 106 tests pass in renderer_test.py and mesh_test.py
including all snapshots; all 16 benchmark gradients are bit-identical to
the pre-split values.
